added weather icon, changed to london weather data

+++ app/src/main/java/io/github/myolwin00/sunshine/data/Weather.kt
@@ -12,7 +12,7 @@ class Weather {
     var mIcon: String? = null
 
     @SerializedName("id")
-    var mId: Long? = null
+    var mId: Int? = null
 
     @SerializedName("main")
     var mMain: String? = null
